rely on correct implicit build rule
Use this standard rule by not specifying any build instructions for
$TARGET: `make -p |& grep "%: %\.c$" -A2`

The previous rule tried to construct a target `build`, which requires
that `$(TARGET)` *is built already* and *afterwards*
invokes the incomplete build command specified in the recipe for
`build`. That's nonsense.

If you want to be more explicit about dependencies,
you could also add another line `$(TARGET): $(TARGET).c`
to make the source explicit but leave the compile command to `make`.
To be independent the implicit rules, you could also add a more complete full rule

```
$(TARGET): $(TARGET).c
	$(CC) $(CFLAGS) $(CPPFLAGS) $(LDFLAGS) $^ %(LDLIBS) -o $@
```

but POSIX specifies

    make shall have a collection of built-in macros and inference rules
    that infer prerequisite relationships to simplify maintenance of programs

so it seems to be safe to assume a fitting rule for c-files.

libs should not be specified in CPPFLAGS because it breaks the parameter
order and can fail with -Wl,--as-needed (debian default).

Also supports LDFLAGS, e.g., to use `-fuse-ld=mold`.

`?=` was not necessary. user-provided variables always overrule simple
assignments, so `make CC=clang` works also without `?`.